[WIP] compile hdf5 kaffe file and run forward pass (#7)
* compiles the znet.so file

* compiles but importing the so file get an error of missing symbol

* fix a avx flag not defined bug

* compiles, but still have the undefined symbol error

* remove the pybind11 path

* rename sergify script and add documentation

* delete tmp cpp files

* fix znnphi path for Makefile

* rename back the znet.h to znet.hpp

* captures the make, not the output path

* the compilation and inference works now!

+++ cpp/include/znn/layer/deconv_as_conv.hpp
@@ -141,7 +141,7 @@ public:
    }
 
    ~DeconvAsConvLayer() {
-      for (int i = 0; i < convs.size(); i++) {
+      for (size_t i = 0; i < convs.size(); i++) {
          delete convs[i];
          delete kernels[i];
       }
@@ -150,7 +150,7 @@ public:
    void forward(float const* __restrict i, float* __restrict o, 
      float const* __restrict runtime_kernel, float const* __restrict runtime_bias, float const* __restrict additive_scale)
    {
-      for (int c = 0; c < convs.size(); c++) {
+      for (size_t c = 0; c < convs.size(); c++) {
          convs[c]->forward(i, o, kernels[c]->data(), runtime_bias, additive_scale); 
       }
    }
